+++ packages/terra-framework-docs/src/terra-dev-site/test/worklist-data-grid/WorklistDataGridWithRowSelection.test.jsx
@@ -41,7 +41,7 @@ const WorklistDataGridWithRowSelection = () => {
   return (
     <WorklistDataGrid
       id="default-terra-worklist-data-grid"
-      columns={cols}
+      overflowColumns={cols}
       rows={[...rows]}
       rowHeaderIndex={rowHeaderIndex}
       ariaLabel="Worklist Data Grid"