+++ packages/terra-worklist-data-grid/src/subcomponents/RowSelectionCell.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import '../_elementPolyfill';
 import { injectIntl } from 'react-intl';
+
 import Cell from './Cell';
 
 const propTypes = {