+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+* Breaking Changes
+  * Removed `columns` prop in favor of the new `pinnedColumns` and `overflowColumns` props.
+
+* Added
+  * Added `pinnedColumns` prop for columns that will be pinned.
+  * Added `overflowColumns` prop for columns that are scrollable.
+  * Added support for pinned columns.
+ 
+* Fixed
+  * Text alignment of the row header column.
+
 ## 0.3.0 - (July 19, 2023)
 
 * Breaking Changes